refactor(generation): reorganize text generation logic into dedicated module
- Moved text generation functions, including `generateText` and `generateObject`, into a new `textGeneration.ts` module, enhancing modularity and separation of concerns.
- Updated import paths in tests and other modules to reflect the new structure, improving code clarity and maintainability.
- Removed deprecated `generateObjectDeprecated` function, streamlining the codebase and promoting adherence to current best practices.
